refactor: refactor calendar structure
- Divided calenda/page.tsx into Header, Days, Menu, and Modal components
- Created hooks directory to improve expandability and readability
- Created types directory to imporve maintainability
- Created utils directory for utility functions
