+++ client/app/types/useMenu.ts
@@ -1,0 +1,7 @@
+export type UseMenuReturnType = {
+  menuOpen: boolean;
+  toggleMenu: () => void;
+  friendListOpen: boolean;
+  toggleFriendList: () => void;
+  menuRef: React.RefObject<HTMLDivElement>;
+};
