+++ client/app/types/useMenuOpen.ts
@@ -1,0 +1,5 @@
+export type UseMenuOpenReturnType = {
+  menuOpen: boolean;
+  setMenuOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  toggleMenu: () => void;
+};
